+++ application/views/common/header.php
@@ -69,6 +69,8 @@
 						</a>
 
 						<ul class="dropdown-menu">
+							<li><a href="<?=base_url('user/favourites')?>">Favourites</a></li>
+							<li class="divider"></li>
 							<li><a href="<?=base_url('user/history')?>">History</a></li>
 							<li class="divider"></li>
 							<li><a href="<?=base_url('user/options')?>">Options</a></li>