+++ public/assets/js/pages/dashboard.js
@@ -246,7 +246,14 @@ $(function(){
 
 			if (--time_left < 0) {
 				clearInterval(timer);
-				//location.reload(); //?
+
+				//Wait one minute, then change favicon to alert user of update
+				setTimeout(function(){
+					//TODO: This "should" just be favicon.updated.ico, and we should handle any ENV stuff on the backend
+					$("link[rel*='icon']").attr("href", base_url+"favicon.production.updated.ico");
+
+					//location.reload(); //TODO: We should have an option for this?
+				}, 60000);
 			}
 		}, 1000);
 	}