+++ public/CHANGELOG.md
@@ -8,7 +8,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [1.3.0] - 2016-10-04
 ### Added
 - Ability to favourite chapters! (This does require the series to be tracked first though!)
-  - There isn't any place to view these yet, maybe soon!
+  - History has also been added for this.
+  - At the moment, the only way to remove a favourite is to go the the chapter page & click favourite again. Will try and add another method soon.
 
 ### Fixed
 - CI garbage collection "should" work properly now.